+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/__init__.py
@@ -4,6 +4,7 @@ from .resources.gcp_resources import gcp_gcs_client, gcs_bucket_manager, gcs_fil
 from .resources.github_resources import github_client, github_connector_repo, github_connectors_directory
 from .resources.local_resources import simple_local_file_manager
 
+from .assets.github_assets import source_controlled_connectors
 from .assets.catalog_assets import (
     oss_destinations_dataframe,
     cloud_destinations_dataframe,
@@ -15,13 +16,16 @@ from .assets.catalog_assets import (
     all_destinations_dataframe,
     connector_catalog_location_markdown,
     connector_catalog_location_html,
-    metadata_definitions,
-    valid_metadata_list,
-    write_metadata_definitions_to_filesystem,
-)
-from .assets.github_assets import source_controlled_connectors
 
-from .jobs.catalog_jobs import generate_catalog_markdown
+)
+from .assets.metadata_assets import (
+    catalog_derived_metadata_definitions,
+    valid_metadata_list,
+)
+
+from .assets.dev_assets import persist_metadata_definitions
+
+from .jobs.catalog_jobs import generate_catalog_markdown, generate_local_metadata_files
 from .sensors.catalog_sensors import catalog_updated_sensor
 
 from .config import REPORT_FOLDER, CATALOG_FOLDER, CONNECTORS_PATH, CONNECTOR_REPO_NAME
@@ -39,9 +43,9 @@ ASSETS = [
     connector_catalog_location_markdown,
     connector_catalog_location_html,
     source_controlled_connectors,
-    metadata_definitions,
+    catalog_derived_metadata_definitions,
     valid_metadata_list,
-    write_metadata_definitions_to_filesystem,
+    persist_metadata_definitions,
 ]
 
 RESOURCES = {
@@ -66,7 +70,7 @@ SENSORS = [catalog_updated_sensor(job=generate_catalog_markdown, resources_def=R
 
 SCHEDULES = []
 
-JOBS = [generate_catalog_markdown]
+JOBS = [generate_catalog_markdown, generate_local_metadata_files]
 
 """
 START HERE
