+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/jobs/catalog_jobs.py
@@ -1,4 +1,5 @@
-from dagster import define_asset_job, job
+from dagster import define_asset_job
 
 
-generate_catalog_markdown = define_asset_job(name="generate_catalog_markdown", selection="*")
+generate_catalog_markdown = define_asset_job(name="generate_catalog_markdown", selection=["connector_catalog_location_html", "connector_catalog_location_markdown"])
+generate_local_metadata_files = define_asset_job(name="generate_local_metadata_files", selection=["persist_metadata_definitions"])
