+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/utils/dagster_helpers.py
@@ -1,0 +1,4 @@
+from dagster import MetadataValue, Output
+
+def OutputDataFrame(result_df):
+    return Output(result_df, metadata={"count": len(result_df), "preview": MetadataValue.md(result_df.to_markdown())})
