--- airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/utils/html.py
@@ -1,12 +1,0 @@
-def html_body(title, content):
-    return f"""
-    <!DOCTYPE html>
-    <html>
-        <head>
-            <title>{title}</title>
-        </head>
-        <body>
-            {content}
-        </body>
-    </html>
-    """
